start implement DEFLATE algo for HALD LUT encode/decode
