+++ LutUtils/include/deflate_algo_utils.h
@@ -1,28 +1,7 @@
-#ifndef __LUT_LIBRARY_DEFLATE_ENCODER_ALGO_UTILS__
-#define __LUT_LIBRARY_DEFLATE_ENCODER_ALGO_UTILS__
+#ifndef __LUT_LIBRARY_DEFLATE_OBJECTS_ALGO_UTILS__
+#define __LUT_LIBRARY_DEFLATE_OBJECTS_ALGO_UTILS__
 
-#include <vector>
+#include "deflate_decoder.h"
+#include "deflate_encoder.h"
 
-class CDeflateBasicDecoder
-{
-public:
-	CDeflateBasicDecoder() { ; }
-	virtual ~CDeflateBasicDecoder() { ; }
-
-	std::vector<uint8_t> Decode (const std::vector<uint8_t>& iDat) {
-		std::vector<uint8_t> dummy{};
-		return dummy;
-	}
-
-private:
-};
-
-
-class CDeflateBasicEncoder
-{
-public:
-private:
-};
-
-
-#endif /* __LUT_LIBRARY_DEFLATE_ENCODER_ALGO_UTILS__ */
+#endif /* __LUT_LIBRARY_DEFLATE_OBJECTS_ALGO_UTILS__ */
