+++ LutUtils/include/deflate_common.h
@@ -1,0 +1,8 @@
+#ifndef __LUT_LIBRARY_DEFLATE_COMMON_ALGO_UTILS__
+#define __LUT_LIBRARY_DEFLATE_COMMON_ALGO_UTILS__
+
+#include <cstdint>
+#include <vector>
+
+
+#endif /* __LUT_LIBRARY_DEFLATE_COMMON_ALGO_UTILS__ */
