+++ LutUtils/src/deflate_decoder.cpp
@@ -1,0 +1,9 @@
+#include "deflate_decoder.h"
+
+
+std::vector<uint8_t> CDeflateBasicDecoder::Decode (const std::vector<uint8_t>& iDat)
+{
+	std::vector<uint8_t> dummy{};
+	return dummy;
+}
+
