+++ LutUtils/src/deflate_encoder.cpp
@@ -1,0 +1,9 @@
+#include "deflate_encoder.h"
+
+
+std::vector<uint8_t> CDeflateBasicEncoder::Encode (const std::vector<uint8_t>& rawData)
+{
+	std::vector<uint8_t> dummy{};
+	return dummy;
+}
+
